requirements.txt
Flask
face_recognition
Pillow
